feat: changed ex10 to sweet_spot
