+++ day00/ex10/ex10.md
@@ -1,35 +1,25 @@
-# Exercise 10 - Rank
+# Exercise 10 - Sweet spot
 
 |                         |                    |
 | -----------------------:| ------------------ |
 |   Turnin directory :    |  ex10              |
-|   Files to turn in :    |  rank.py           |
+|   Files to turn in :    |  sweet_spot.py     |
 |   Forbidden function :  |  None              |
 |   Remarks :             |  n/a               |
-| Allowed SQL Keywords :  |  all               |
-
+| Allowed SQL Keywords :  | SELECT, ORDER BY, GROUP BY, COUNT, date_trunc |
 
 ## Objective
 
-Learn to rank your data based on calculated values.
+Find the month where the most important number of games are sold.
 
 ## Instructions
 
-Turn in the top 100 games that have the best average rating, and the top 100 games that have the worst average rating. Names of games must be sorted alphabetically.
+Find the month where the most important number of games are sold.
 
-**You must only use PostgreSQL for your queries!**
+**You can use PostgreSQL and Python**
 
 # Example
 
 ```
-100 - Cups and Balls Ed. HD
-100 Falling Balls - Fill The Cups: mini arcade game edition
-100 Pool Balls - Billiard Ed.
-100 - The Edition of Cups and Balls
-...
-Bigfoot Hunter : Deadly Shores Of Challenge Swat
-Block Adventure World - Zombie Game Edition
-Bobby Fischer Complete Collection
-Brick By Brick Physics Game
-Bricks of Rome
+sweet_spot : january
 ```
